Fixed the unit testing files to only test inside the browser. #19
